Move tooltips to status tags
